refactor(global): :zap: Navbar become in a shared component
