+++ apps/suite/components/index.tsx
@@ -1,7 +1,6 @@
 import { AppWrapper } from './AppWrapper';
 import { Footer } from './Footer';
-import { Navbar } from './Navbar';
 import { Sidebar } from './Sidebar';
 import { SWRProvider } from './SWRProvider';
 
-export { Navbar, Footer, Sidebar, AppWrapper, SWRProvider };
+export { Footer, Sidebar, AppWrapper, SWRProvider };
