+++ packages/ui/src/index.tsx
@@ -12,6 +12,7 @@ export { FormWrapper, type FormWrapperProps } from './components/FormWrapper';
 export { InputField, type InputFieldProps } from './components/InputField';
 export { InputGroup, type InputProps } from './components/InputGroup';
 export { Modal, type ModalProps } from './components/Modal';
+export { Navbar, type NavbarProps } from './components/Navbar';
 export { SelectField, type SelectFieldProps } from './components/SelectField';
 export { ThemeSwitcher } from './components/ThemeSwitcher';
 //ASSETS
